refactor: rename add-env.ts to env.ts, disabled env import test
Renamed test helper file to better reflect it's content. Seperated and commented out the env import
test for now. Removed an extra carriage return in deleteProject
